deleted turbomobilesendtoend js file
